+++ packages/eslint-config-fbjs/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+### Added
+- `fbjs/strict` config, with the same rules enabled as the default config, with each warning upgraded to an error.
+
 ### Changed
 - Switched to use eslint-plugin-flowtype for Flow-related rules